+++ runtime/src/lib.rs
@@ -1172,7 +1172,6 @@ where
             frame_system::CheckNonce::<Runtime>::from(nonce),
             frame_system::CheckWeight::<Runtime>::new(),
             pallet_transaction_payment::ChargeTransactionPayment::<Runtime>::from(tip),
-            pallet_free_tx::ValidateElectricityPayment::<Runtime>::new(),
             pallet_registry::ValidateConfirmRegistered::<Runtime>::new(),
         );
         let raw_payload = SignedPayload::new(call, extra)
@@ -1413,16 +1412,6 @@ impl pallet_transaction_pause::Config for Runtime {
     type WeightInfo = weights::pallet_transaction_pause::WeightInfo<Runtime>;
 }
 
-impl pallet_free_tx::Config for Runtime {
-    type RuntimeCall = RuntimeCall;
-    type RuntimeEvent = RuntimeEvent;
-    type UpdateOrigin = EitherOfDiverse<
-        EnsureRoot<AccountId>,
-        pallet_collective::EnsureProportionAtLeast<AccountId, CouncilCollective, 2, 3>,
-    >;
-    type WeightInfo = weights::pallet_free_tx::WeightInfo<Runtime>;
-}
-
 impl pallet_propagation::Config for Runtime {
     type RuntimeEvent = RuntimeEvent;
 }
@@ -1477,8 +1466,7 @@ construct_runtime!(
     Slashing: pallet_slashing = 52,
     Programs: pallet_programs = 53,
     TransactionPause: pallet_transaction_pause = 54,
-    FreeTx: pallet_free_tx = 55,
-    Propagation: pallet_propagation = 56,
+    Propagation: pallet_propagation = 55,
 
 
   }
@@ -1508,7 +1496,6 @@ pub type SignedExtra = (
     frame_system::CheckNonce<Runtime>,
     frame_system::CheckWeight<Runtime>,
     pallet_transaction_payment::ChargeTransactionPayment<Runtime>,
-    pallet_free_tx::ValidateElectricityPayment<Runtime>,
     pallet_registry::ValidateConfirmRegistered<Runtime>,
 );
 /// Unchecked extrinsic type as expected by this runtime.
@@ -1549,7 +1536,6 @@ mod benches {
       [pallet_election_provider_multi_phase, ElectionProviderMultiPhase]
       [frame_election_provider_support, EPSBench::<Runtime>]
       [pallet_elections_phragmen, Elections]
-      [pallet_free_tx, FreeTx]
       [pallet_staking_extension, StakingExtension]
       [pallet_grandpa, Grandpa]
       [pallet_im_online, ImOnline]