+++ runtime/src/weights/mod.rs
@@ -37,7 +37,6 @@ pub mod pallet_collective;
 pub mod pallet_democracy;
 pub mod pallet_election_provider_multi_phase;
 pub mod pallet_elections_phragmen;
-pub mod pallet_free_tx;
 pub mod pallet_identity;
 pub mod pallet_im_online;
 pub mod pallet_indices;